created logsong
